requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.18.0
requests==2.25.1
soupsieve==2.2.1
SQLAlchemy==1.4.20
urllib3==1.26.6
Werkzeug==2.0.1
